requirements.txt
SQLAlchemy==2.0.37
fastapi==0.115.2
uvicorn==0.32.0
alembic==1.13.3
python-dotenv==1.0.1
httpx==0.28.1
